Agregar el campo de riesgoso en la tabla de estado de deserciones.

+++ app/controllers/super_user_controller.rb
@@ -8,12 +8,12 @@ class SuperUserController < ApplicationController
 
 	def new_estados_desercion
 		# estados_desercion = EstadoDesercion.select([:id, :nombre_estado, :notificar]).order(nombre_estado: :asc)
-		estados_desercion = EstadoDesercion.select([:id, :nombre_estado, :notificar]).order(:nombre_estado => :asc)
+		estados_desercion = EstadoDesercion.getEstados
 		render action: :index, locals: {partial: 'new_estado_desercion', estado_desercion: EstadoDesercion.new, estados: estados_desercion, context: 'estado'}
 	end
 
 	def modify_estados_desercion
-		estados_desercion = EstadoDesercion.select([:id, :nombre_estado, :notificar]).order(:nombre_estado => :asc)
+		estados_desercion = EstadoDesercion.getEstados
 		render action: :index, locals: {partial: 'modify_estado_desercion', context: 'estado', estados: estados_desercion}
 		
 	end
@@ -44,35 +44,42 @@ class SuperUserController < ApplicationController
 		ed_obj = EstadoDesercion.find_by(id: estado_params[:id])
 
 		if !ed_obj.nil?
-			if params[:to_delete] == "1"
-				# Borrar estado (REVISAR SI HAY QYE VERIFICAR SI HAY ESTUDIANTES CON ESE ESTADO ANTES DE BORRAR).
-				ed_obj.destroy!
-				estados_desercion = EstadoDesercion.select([:id, :nombre_estado, :notificar]).order(:nombre_estado => :asc)
+			# Verificar que el estado no sea fijo.
+			if ed_obj.nombre_estado != EstadoDesercion::DESERTO_ESTADO
+				if params[:to_delete] == "1"
+					# Borrar estado (REVISAR SI HAY QYE VERIFICAR SI HAY ESTUDIANTES CON ESE ESTADO ANTES DE BORRAR).
+					ed_obj.destroy!
+					estados_desercion = EstadoDesercion.getEstados
 
-				render json: {msg: "Estado de deserción actualizado exitosamente.", table: render_to_string(partial: 'estados_desercion_table', formats: [:html], layout: false, locals: {estados: estados_desercion}), type: "success"}
+					render json: {msg: "Estado de deserción actualizado exitosamente.", table: render_to_string(partial: 'estados_desercion_table', formats: [:html], layout: false, locals: {estados: estados_desercion}), type: "success"}
 
-			elsif params[:to_delete] == "0"
-				# Actualizar estado.
-				ed_obj.nombre_estado = estado_params[:nombre_estado]
-				ed_obj.notificar = estado_params[:notificar]
+				elsif params[:row_edited] == "1"
+					# Actualizar estado.
+					ed_obj.nombre_estado = estado_params[:nombre_estado]
+					ed_obj.notificar = estado_params[:notificar]
+					ed_obj.riesgoso = estado_params[:riesgoso]
 
-				# Verificar que el nombre nuevo del estado no exista en la BD.
-				if !ed_obj.getEstadoDesercion(ed_obj.id)
-					if ed_obj.save
-						estados_desercion = EstadoDesercion.select([:id, :nombre_estado, :notificar]).order(:nombre_estado => :asc)
+					# Verificar que el nombre nuevo del estado no exista en la BD.
+					if !ed_obj.getEstadoDesercion(ed_obj.id)
+						if ed_obj.save
+							estados_desercion = EstadoDesercion.getEstados
 
-						render json: {msg: "Estado de deserción actualizado exitosamente.", table: render_to_string(partial: 'estados_desercion_table', formats: [:html], layout: false, locals: {estados: estados_desercion}), type: "success", estado_obj: ed_obj}
+							render json: {msg: "Estado de deserción actualizado exitosamente.", table: render_to_string(partial: 'estados_desercion_table', formats: [:html], layout: false, locals: {estados: estados_desercion}), type: "success", estado_obj: ed_obj}
 
+						else
+							render json: {msg: ed_obj.getFormatErrorMessages, type: 'danger'}, status: 422
+						end
 					else
-						render json: {msg: ed_obj.getFormatErrorMessages, type: 'danger'}, status: 422
+						render json: {msg: "El estado de deserción ya se encuentra ingresado.", type: 'danger'}, status: 422
+
 					end
 				else
-					render json: {msg: "El estado de deserción ya se encuentra ingresado.", type: 'danger'}, status: 422
-
+					# Error de opcion.
+					render json: {msg: "Ha ocurrido un error con la opción ingresada.", type: 'danger'}, status: 422
 				end
 			else
-				# Error de opcion.
-				render json: {msg: "Ha ocurrido un error con la opción ingresada.", type: 'danger'}, status: 422
+				# Es el estado de desercion fijo,
+				render json: {msg: "El estado de desercion selecionado no se puede borrar ni modificar.", type: "danger"}, status: :bad_request
 			end
 		else
 			render json: {msg: "Ha ocurrido un problema con encontrar el estado de desercion en el sistema.", type: "danger"}, status: 422
@@ -159,7 +166,7 @@ class SuperUserController < ApplicationController
 	end
 
 	def estado_desercion_params
-		params.require(:estado_desercion).permit(:id, :nombre_estado, :notificar)
+		params.require(:estado_desercion).permit(:id, :nombre_estado, :notificar, :riesgoso)
 	end
 
 	def sign_up_params
